requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alpha-vantage==2.3.1
async-timeout==4.0.2
attrs==22.2.0
avanza-api==7.1.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.9.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
discord.py==2.2.2
Flask==2.2.3
fonttools==4.39.2
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
Pillow==9.4.0
pip==22.0.2
pip-autoremove==0.10.0
platformdirs==3.2.0
plotly==5.13.1
pyotp==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
setuptools==59.6.0
six==1.16.0
ta==0.10.2
tenacity==8.2.2
tomli==2.0.1
urllib3==1.26.15
websockets==10.4
Werkzeug==2.2.3
yarl==1.8.2
